Move Local structure definition to localTemplate (#3161)
Move Local structure definition to localTemplate, so that changes to localTemplate are more obvious to be a change of template than a change of data.

Tested manually.
