+++ config/defaultsGenerator/defaultsGenerator.go
@@ -39,7 +39,7 @@ var jsonExampleFileName = flag.String("j", "", "Name of the json example file")
 
 var autoGenHeader = `
 // This file was auto generated by ./config/defaultsGenerator/defaultsGenerator.go, and SHOULD NOT BE MODIFIED in any way
-// If you want to make changes to this file, make the corresponding changes to Local in config.go and run "go generate".
+// If you want to make changes to this file, make the corresponding changes to Local in localTemplate.go and run "go generate".
 `
 
 // printExit prints the given formatted string ( i.e. just like fmt.Printf ), with the defaultGenerator executable program name
