+++ installer/config.json.example
@@ -44,8 +44,10 @@
     "EnableProfiler": false,
     "EnableRequestLogger": false,
     "EnableTopAccountsReporting": false,
+    "EnableVerbosedTransactionSyncLogging": false,
     "EndpointAddress": "127.0.0.1:0",
     "FallbackDNSResolverAddress": "",
+    "ForceFetchTransactions": false,
     "ForceRelayMessages": false,
     "GossipFanout": 4,
     "IncomingConnectionsLimit": 800,
@@ -69,7 +71,7 @@
     "OutgoingMessageFilterBucketSize": 128,
     "ParticipationKeysRefreshInterval": 60000000000,
     "PeerConnectionsUpdateInterval": 3600,
-    "PeerPingPeriodSeconds": 10,
+    "PeerPingPeriodSeconds": 0,
     "PriorityPeers": {},
     "PublicAddress": "",
     "ReconnectTime": 60000000000,
@@ -82,6 +84,8 @@
     "TLSCertFile": "",
     "TLSKeyFile": "",
     "TelemetryToLog": true,
+    "TransactionSyncDataExchangeRate": 0,
+    "TransactionSyncSignificantMessageThreshold": 0,
     "TxPoolExponentialIncreaseFactor": 2,
     "TxPoolSize": 15000,
     "TxSyncIntervalSeconds": 60,
