+++ test/testdata/configs/config-v17.json
@@ -48,7 +48,7 @@
     "FallbackDNSResolverAddress": "",
     "ForceRelayMessages": false,
     "GossipFanout": 4,
-    "IncomingConnectionsLimit": 10000,
+    "IncomingConnectionsLimit": 800,
     "IncomingMessageFilterBucketCount": 5,
     "IncomingMessageFilterBucketSize": 512,
     "IsIndexerActive": false,
