+++ test/testdata/configs/config-v18.json
@@ -71,7 +71,7 @@
     "OutgoingMessageFilterBucketSize": 128,
     "ParticipationKeysRefreshInterval": 60000000000,
     "PeerConnectionsUpdateInterval": 3600,
-    "PeerPingPeriodSeconds": 10,
+    "PeerPingPeriodSeconds": 0,
     "PriorityPeers": {},
     "PublicAddress": "",
     "ReconnectTime": 60000000000,
